Improve agent switching, plugin lifecycle, and Agent tab streaming performance (#321)
Wire Settings agent selection through persisted config with explicit Apply/restart,
harden plugin install/remove/reinstall flows, and forward repo .env to agent PTY
sessions. Optimize the Agent panel to stream PTY output without idle repaints or
per-frame label churn, and document the pipeline for future tool authors.

+++ plugins/kiwi_mcp_context/src/mcp.rs
@@ -191,7 +191,7 @@ fn dispatch_search(
     let limit = args
         .get("limit")
         .and_then(|v| v.as_i64())
-        .unwrap_or(8) as i32;
+        .unwrap_or(8);
     let session_key = args
         .get("session_key")
         .and_then(|v| v.as_str())
@@ -220,7 +220,7 @@ fn dispatch_list(args: &Value, db: &mut ContextDb) -> anyhow::Result<String> {
     let limit = args
         .get("limit")
         .and_then(|v| v.as_i64())
-        .unwrap_or(20) as i32;
+        .unwrap_or(20);
     let session_key = args
         .get("session_key")
         .and_then(|v| v.as_str())
